+++ CHANGELOG.md
@@ -10,4 +10,6 @@
 
 - remove redundant exception classes;
 
-- add tree sorter interface.
+- add tree sorter interface;
+
+- remove "_" from transliterator's allowed symbol list.